core: address PR review feedback on the tick timeout
Three reviewer nits, no behaviour change (ch572-uart image stays byte-identical,
330 host tests green):

- OB_IDLE_TIMEOUT_TICKS: widen the ms*ticks-per-ms product to uint64 before the
  uint32 cast, matching the static assert's own 64-bit intermediate. The 2^31
  cap already makes the 32-bit product safe; this makes the intent explicit and
  future-proofs a knob/clock change.
- ob_idle_elapsed(): name the parameters *_ticks in the C signature and the
  ctypes wrapper, so the unit is on the surface rather than in the doc only.
- boot_decision.h: state the caller's wrap-safety obligation on ob_idle_elapsed
  itself (bound timeout_ticks below 2^32 with margin AND sample often enough not
  to skip the [timeout, 2^32) window), since it is the reusable primitive.

+++ firmware/core/main.c
@@ -38,7 +38,7 @@
  * 60 MHz, ~335 s at 6.4 MHz; the 10 s default and any realistic idle window sit
  * far inside it (see firmware/README.md). */
 #define OB_IDLE_TIMEOUT_TICKS \
-    ((uint32_t)((OB_IDLE_TIMEOUT_MS) * (OB_CPU_HZ / 1000u)))
+    ((uint32_t)((uint64_t)(OB_IDLE_TIMEOUT_MS) * (OB_CPU_HZ / 1000u)))
 _Static_assert(OB_CPU_HZ % 1000u == 0u, "OB_CPU_HZ must be a whole number of kHz");
 _Static_assert(OB_IDLE_TIMEOUT_MS == 0u ||
                (uint64_t)(OB_IDLE_TIMEOUT_MS) * (OB_CPU_HZ / 1000u) <= 0x80000000u,
